Move snapshot builders to utils and rename to parsers
