+++ server/__tests__/suites/integration/players.test.ts
@@ -4,7 +4,7 @@ import MockAdapter from 'axios-mock-adapter';
 import prisma from '../../../src/prisma';
 import env from '../../../src/env';
 import apiServer from '../../../src/api';
-import { BOSSES, Metric, PlayerStatus, PlayerType, SnapshotDataSource } from '../../../src/utils';
+import { BOSSES, Metric, PlayerStatus, PlayerType } from '../../../src/utils';
 import {
   registerCMLMock,
   registerHiscoresMock,
@@ -14,7 +14,6 @@ import {
   sleep,
   resetRedis
 } from '../../utils';
-import * as snapshotServices from '../../../src/api/modules/snapshots/snapshot.services';
 import * as snapshotUtils from '../../../src/api/modules/snapshots/snapshot.utils';
 import * as playerServices from '../../../src/api/modules/players/player.services';
 import * as playerEvents from '../../../src/api/modules/players/player.events';
@@ -1466,11 +1465,10 @@ describe('Player API', () => {
 
       const player = playerResponse.body;
 
-      const previousSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: globalData.hiscoresRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const previousSnapshot = await snapshotUtils.parseHiscoresSnapshot(
+        player.id,
+        globalData.hiscoresRawData
+      );
 
       const modifiedRawData = modifyRawHiscoresData(globalData.hiscoresRawData, [
         { metric: Metric.RUNECRAFTING, value: 50_000_000 },
@@ -1483,11 +1481,7 @@ describe('Player API', () => {
         [PlayerType.IRONMAN]: { statusCode: 404 }
       });
 
-      const rejectedSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: modifiedRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const rejectedSnapshot = await snapshotUtils.parseHiscoresSnapshot(player.id, modifiedRawData);
 
       // Manually set overall ranks and exp for de-iron checks later on
       previousSnapshot.overallRank = 10_000;
@@ -1597,11 +1591,10 @@ describe('Player API', () => {
 
       const player = playerResponse.body;
 
-      const previousSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: globalData.hiscoresRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const previousSnapshot = await snapshotUtils.parseHiscoresSnapshot(
+        player.id,
+        globalData.hiscoresRawData
+      );
 
       const modifiedRejectedRawData = modifyRawHiscoresData(globalData.hiscoresRawData, [
         { metric: Metric.ZULRAH, value: 1615 }, // zulrah kc dropped from 1646 to 1615
@@ -1613,11 +1606,7 @@ describe('Player API', () => {
         [PlayerType.IRONMAN]: { statusCode: 404 }
       });
 
-      const rejectedSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: modifiedRejectedRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const rejectedSnapshot = await snapshotUtils.parseHiscoresSnapshot(player.id, modifiedRejectedRawData);
 
       const flagContext = reviewFlaggedPlayer(player, previousSnapshot, rejectedSnapshot);
 
@@ -1690,11 +1679,10 @@ describe('Player API', () => {
 
       const player = playerResponse.body;
 
-      const previousSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: globalData.hiscoresRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const previousSnapshot = await snapshotUtils.parseHiscoresSnapshot(
+        player.id,
+        globalData.hiscoresRawData
+      );
 
       await sleep(500);
 
@@ -1711,11 +1699,7 @@ describe('Player API', () => {
         [PlayerType.IRONMAN]: { statusCode: 404 }
       });
 
-      const rejectedSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: modifiedRejectedRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const rejectedSnapshot = await snapshotUtils.parseHiscoresSnapshot(player.id, modifiedRejectedRawData);
 
       const flagContext = reviewFlaggedPlayer(player, previousSnapshot, rejectedSnapshot);
       expect(flagContext).toBeNull();
@@ -1754,11 +1738,7 @@ describe('Player API', () => {
 
       const player = playerResponse.body;
 
-      const previousSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: modifiedPreviousRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const previousSnapshot = await snapshotUtils.parseHiscoresSnapshot(player.id, modifiedPreviousRawData);
 
       await sleep(500);
 
@@ -1767,11 +1747,10 @@ describe('Player API', () => {
         [PlayerType.IRONMAN]: { statusCode: 404 }
       });
 
-      const rejectedSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: globalData.hiscoresRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const rejectedSnapshot = await snapshotUtils.parseHiscoresSnapshot(
+        player.id,
+        globalData.hiscoresRawData
+      );
 
       const flagContext = reviewFlaggedPlayer(player, previousSnapshot, rejectedSnapshot);
       expect(flagContext).toBeNull();
@@ -1809,11 +1788,10 @@ describe('Player API', () => {
       // pre changes setup
       const groupId = await setupPreTransitionData(1000, player.id);
 
-      const previousSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: globalData.hiscoresRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const previousSnapshot = await snapshotUtils.parseHiscoresSnapshot(
+        player.id,
+        globalData.hiscoresRawData
+      );
 
       await prisma.snapshot.create({ data: previousSnapshot });
 
@@ -1832,11 +1810,7 @@ describe('Player API', () => {
         [PlayerType.IRONMAN]: { statusCode: 404 }
       });
 
-      const rejectedSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: modifiedRejectedRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const rejectedSnapshot = await snapshotUtils.parseHiscoresSnapshot(player.id, modifiedRejectedRawData);
 
       const submitNameChangeResponse = await api.post(`/names`).send({
         oldName: 'Greg Hirsch',
@@ -2005,11 +1979,10 @@ describe('Player API', () => {
       // pre changes setup
       const groupId = await setupPreTransitionData(2000, player.id);
 
-      const previousSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: globalData.hiscoresRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const previousSnapshot = await snapshotUtils.parseHiscoresSnapshot(
+        player.id,
+        globalData.hiscoresRawData
+      );
 
       await prisma.snapshot.create({ data: previousSnapshot });
 
@@ -2028,11 +2001,7 @@ describe('Player API', () => {
         [PlayerType.IRONMAN]: { statusCode: 404 }
       });
 
-      const rejectedSnapshot = await snapshotServices.buildSnapshot({
-        playerId: player.id,
-        rawCSV: modifiedRejectedRawData,
-        source: SnapshotDataSource.HISCORES
-      });
+      const rejectedSnapshot = await snapshotUtils.parseHiscoresSnapshot(player.id, modifiedRejectedRawData);
 
       const submitNameChangeResponse = await api.post(`/names`).send({
         oldName: 'TomWambsgans',
