+++ server/src/api/modules/players/services/UpdatePlayerService.ts
@@ -208,7 +208,7 @@ async function fetchStats(player: Player, type?: PlayerType): Promise<Snapshot> 
   const hiscoresCSV = await jagexService.fetchHiscoresData(player.username, type || player.type);
 
   // Convert the csv data to a Snapshot instance
-  const newSnapshot = await snapshotServices.buildSnapshot({ playerId: player.id, rawCSV: hiscoresCSV });
+  const newSnapshot = await snapshotUtils.parseHiscoresSnapshot(player.id, hiscoresCSV);
 
   return newSnapshot;
 }
