+++ server/src/api/modules/snapshots/snapshot.services.ts
@@ -1,4 +1,3 @@
-export * from './services/BuildSnapshotService';
 export * from './services/RollbackSnapshotsService';
 export * from './services/FindGroupSnapshotsService';
 export * from './services/FindPlayerSnapshotService';
