+++ server/src/api/modules/snapshots/snapshot.utils.ts
@@ -1,3 +1,4 @@
+import csv from 'csvtojson';
 import {
   BOSSES,
   getMetricRankKey,
@@ -38,6 +39,99 @@ import {
   SkillValue,
   SkillValueWithPlayer
 } from './snapshot.types';
+
+// Skip Deadman Points and Legacy Bounty Hunter (hunter/rogue)
+export const SKIPPED_ACTIVITY_INDICES = [1, 4, 5];
+
+async function parseHiscoresSnapshot(playerId: number, rawCSV: string) {
+  // Convert the CSV text into an array of values
+  // Ex: for skills, each row is [rank, level, experience]
+  const rows = await csv({ noheader: true, output: 'csv' }).fromString(rawCSV);
+
+  // If a new skill/activity/boss was added to the hiscores,
+  // prevent any further snapshot saves to prevent incorrect DB data
+  if (rows.length !== SKILLS.length + ACTIVITIES.length + BOSSES.length + SKIPPED_ACTIVITY_INDICES.length) {
+    throw new ServerError('The OSRS Hiscores were updated. Please wait for a fix.');
+  }
+
+  const snapshotFields = {
+    playerId,
+    createdAt: new Date()
+  };
+
+  let totalExp = 0;
+
+  // Populate the skills' values with the values from the csv
+  SKILLS.forEach((s, i) => {
+    const [rank, , experience] = rows[i];
+    const expNum = parseInt(experience);
+
+    snapshotFields[getMetricRankKey(s)] = parseInt(rank);
+    snapshotFields[getMetricValueKey(s)] = s === Metric.OVERALL && expNum === 0 ? -1 : expNum;
+
+    if (s !== Metric.OVERALL) totalExp += Math.max(0, expNum);
+  });
+
+  // If this player is unranked in overall exp, we should set their overall exp to the total exp of all skills
+  // since that's at least closer to the real number than -1
+  if (snapshotFields[getMetricValueKey(Metric.OVERALL)] === -1 && totalExp > 0) {
+    snapshotFields[getMetricValueKey(Metric.OVERALL)] = totalExp;
+  }
+
+  // Populate the activities' values with the values from the csv
+  for (let i = 0; i < ACTIVITIES.length + SKIPPED_ACTIVITY_INDICES.length; i++) {
+    if (SKIPPED_ACTIVITY_INDICES.includes(i)) continue;
+
+    const [rank, score] = rows[SKILLS.length + i];
+    const skippedCount = SKIPPED_ACTIVITY_INDICES.filter(x => x < i).length;
+
+    const activity = ACTIVITIES[i - skippedCount];
+
+    snapshotFields[getMetricRankKey(activity)] = parseInt(rank);
+    snapshotFields[getMetricValueKey(activity)] = parseInt(score);
+  }
+
+  // Populate the bosses' values with the values from the csv
+  BOSSES.forEach((s, i) => {
+    const [rank, kills] = rows[SKILLS.length + ACTIVITIES.length + SKIPPED_ACTIVITY_INDICES.length + i];
+
+    snapshotFields[getMetricRankKey(s)] = parseInt(rank);
+    snapshotFields[getMetricValueKey(s)] = parseInt(kills);
+  });
+
+  return snapshotFields as Snapshot;
+}
+
+async function parseCMLSnapshot(playerId: number, rawCSV: string) {
+  // CML separates the data "blocks" by a space, for whatever reason.
+  // These blocks are the datapoint timestamp, and the experience and rank arrays respectively.
+  const rows = rawCSV.split(' ');
+  const [timestamp, experienceCSV, ranksCSV] = rows;
+
+  // Convert the experience and rank from CSV data into arrays
+  const exps = (await csv({ noheader: true, output: 'csv' }).fromString(experienceCSV))[0];
+  const ranks = (await csv({ noheader: true, output: 'csv' }).fromString(ranksCSV))[0];
+
+  // If a new skill/activity/boss was added to the CML API,
+  // prevent any further snapshot saves to prevent incorrect DB data
+  if (exps.length !== SKILLS.length || ranks.length !== SKILLS.length) {
+    throw new ServerError('The CML API was updated. Please wait for a fix.');
+  }
+
+  const snapshotFields = {
+    playerId,
+    importedAt: new Date(),
+    createdAt: new Date(parseInt(timestamp, 10) * 1000) // CML stores timestamps in seconds, we need milliseconds
+  };
+
+  // Populate the skills' values with experience and rank data
+  SKILLS.forEach((s, i) => {
+    snapshotFields[getMetricRankKey(s)] = parseInt(ranks[i]);
+    snapshotFields[getMetricValueKey(s)] = parseInt(exps[i]);
+  });
+
+  return snapshotFields as Snapshot;
+}
 
 // On this date, the Bounty Hunter was updated and scores were reset.
 const BOUNTY_HUNTER_UPDATE_DATE = new Date('2023-05-24T10:30:00.000Z');
@@ -417,6 +511,8 @@ function isZerker(snapshot: Snapshot) {
 }
 
 export {
+  parseCMLSnapshot,
+  parseHiscoresSnapshot,
   format,
   average,
   hasChanged,
